CKT

Description

Circuis REST API Communication

Documentation

data Circuit

Available Circuits

Constructors

OP1
Miller Amplifier (op1)

OP2
Symmetrical Amplifier (op2)

OP8
Folded Cascode (op8)

data PDK

Available PDKs

Constructors

XH035
X-Fab XH035 350nm Process

XH018
X-Fab XH018 180nm Process

data Space

Available Design / Action Spaces

Constructors

Electric
Electric Design Space

Geometric
Geometric Design Space

type CircusUrl = String

Base Route to Cricus Server

url :: String -> String -> String -> String -> String -> String -> CircusUrl

Generate URL to a Circuit server from meta information

numEnvs :: CircusUrl -> IO Int

Get the number of Environments from the given Circus Server instance

actionSpace :: CircusUrl -> IO Int

Get the number of Environments from the given Circus Server instance

observationSpace :: CircusUrl -> IO (Int, Int, Int)

Get the number of Environments from the given Circus Server instance

reset :: CircusUrl -> IO (Tensor, Tensor, Tensor)

Reset Environment at given URL reset'' :: CircusUrl -> [Bool] -> IO (Observation [[Float]]) reset'' addr [] = do res <- get addr "reset" let dec = decodeStrict res :: Maybe (Observation [[Float]]) when (isNothing dec) do print res pure $ fromJust dec reset'' addr mask = do res <- post addr "reset" msk let dec = decodeStrict res :: Maybe (Observation [[Float]]) when (isNothing dec) do print res pure $ fromJust dec where msk = toJSON (M.fromList [("env_mask", mask)] :: (M.Map String [Bool]))

Reset Environments and get (observation, achieved_goal, desired_goal)

reset' :: CircusUrl -> Tensor -> IO (Tensor, Tensor, Tensor)

Reset subset of environments with given mask

step :: CircusUrl -> Tensor -> IO (Tensor, Tensor, Tensor, Tensor, Tensor)

Shorthand for taking a Tensor action and returning Tensors

randomAction :: CircusUrl -> IO Tensor

Shorthand for getting Tensors

calculateReward :: CircusUrl -> Tensor -> Tensor -> IO Tensor

Shorthand for calculating reward given Tensors

data Observation a

Observation returned from Stepping / Resetting Environment

Constructors

Observation

Fields

observation :: !a

State

achievedGoal :: !a

Achieved Goal

desiredGoal :: !a

Desired Goal

done :: Maybe [Bool]

Terminal Flag

reward :: Maybe [Float]

Rewards

info :: Maybe [Info]

Info Dict

data Info

Info Dict returned from Environment

Constructors

Info

Fields

goal :: ![String]

Goal Parameters

inputs :: ![String]

Parameters in Action Vector

outputs :: ![String]

Parameters in State Vector

newtype Action a

Action passed to Environment

Constructors

Action

Fields

action :: a

currentPerformance :: CircusUrl -> IO (Map Int (Map String Float))

Shorthand for getting Performance

currentGoal :: CircusUrl -> IO (Map Int (Map String Float))

Shorthand for getting Goal

currentSizing :: CircusUrl -> IO (Map Int (Map String Float))

Shorthand for getting Sizing

lastAction :: CircusUrl -> IO (Map Int (Map String Float))

Shorthand for getting last Action