Lib

Description

Utility and Helper functions for EDELWACE

Documentation

data Algorithm

Available Algorithms

Constructors

TD3
Twin Delayed Deep Deterministic Policy Gradient

SAC
Soft Actor Critic

PPO
Proximal Policy Optimization

data ReplayMemory

Available Replay Buffer Types

Constructors

RPB
Vanilla Replay Buffer

PER
Prioritized Experience Replay

MEM
PPO Style replay Memory

ERE
Emphasizing Recent Experience

HER
Hindsight Experience Replay

data Strategy

Hindsight Experience Replay Strategies for choosing Goals

Constructors

Final
Only Final States are additional targets

Random
Replay with k random states encountered so far (basically vanilla)

Episode
Replay with k random states from same episode.

Future
Replay with k random states from same episode, that were observed after

data Mode

Run Mode

Constructors

Train
Start Agent Training

Cont
Continue Agent Training

Eval
Evaluate Agent

data Args

Command Line Arguments

Constructors

Args

Fields

cktHost :: String

Circus Server Host Address

cktPort :: String

Circus Server Port

cktId :: String

ACE Circuit ID

cktBackend :: String

ACE Backend / PDK

cktSpace :: String

Design Space

algorithm :: String

RL Algorithm

buffer :: String

Replay Buffer

path :: String

Checkpoint Base Path

mlfHost :: String

MLFlow Server Host Address

mlfPort :: String

MLFlow Server Port

mode :: String

Run Mode

config :: String

File Path to config.yaml

type Transition = (Tensor, Tensor, Tensor, Tensor, Tensor)

Type Alias for Transition Tuple (state, action, reward, state', done)

successRate :: Tensor -> Float

Calculate success rate given 1D Boolean done Tensor

range :: Int -> [Int]

Range from 0 to n - 1

fst' :: (a, b, c) -> a

First of triple

delete' :: Ord a => [a] -> Set a -> Set a

Delete multiple elements from a Set

splits :: [Int] -> [[Int]]

Helper function creating split indices

fst3 :: (a, b, c) -> a

First of Triple

both :: (a -> b) -> (a, a) -> (b, b)

Apply a function to both elements of a Tuple

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

Uncurry Triple

lookup' :: Ord k => [k] -> Map k a -> Maybe [a]

Like lookup but for a list of keys.

tmap :: (Tensor -> Tensor) -> Transition -> Transition

Map an appropriate function over a transition tuple

(//) :: Integral a => a -> a -> a

Infix div

(%) :: Integral a => a -> a -> a

Infix mod

currentTimeStamp :: String -> IO String

Current Timestamp as formatted string

currentTimeStamp' :: IO String

Current Timestamp with default formatting: "%Y%m%d-%H%M%S"

createModelArchiveDir :: String -> IO String

Create a model archive directory for the given algorithm

createModelArchiveDir' :: String -> String -> String -> String -> String -> String -> IO String

Create a model archive directory for the given algorithm, ace id and backend

saveOptim :: Adam -> FilePath -> IO ()

Optimizer moments at given prefix

loadOptim :: Int -> Float -> Float -> FilePath -> IO Adam

Load Optimizer State

weightLimit :: Linear -> Float

Calculate weight Limits based on Layer Dimensions

data Initializer

Type of weight initialization

Constructors

Normal
Normally distributed weights

Uniform
Uniformally distributed weights

XavierNormal
Using T.xavierNormal

XavierUniform
Using T.xavierUniform

KaimingNormal
Using T.kaimingNormal

KaimingUniform
Using T.kaimingUniform

Dirac

Eye

Ones

Zeros

Constant

initWeights :: Initializer -> Float -> Float -> [Int] -> IO IndependentTensor

Weights for a layer given limits and dimensions.

weightInit :: Initializer -> Float -> Float -> Linear -> IO Linear

Initialize Weights of Linear Layer

weightInit' :: Initializer -> Float -> Float -> Linear -> IO Linear

Initialize Weights and Bias of Linear Layer

weightInitUniform :: Float -> Float -> Linear -> IO Linear

Initialize weights uniformally given upper and lower bounds

weightInitUniform' :: Linear -> IO Linear

Initialize weights uniformally based on Fan In

weightInitNormal :: Float -> Float -> Linear -> IO Linear

Initialize weights normally given mean and std bounds

weightInitNormal' :: Linear -> IO Linear

Initialize weights normally based on Fan In

softUpdate :: Tensor -> Tensor -> Tensor -> Tensor

Softly update parameters from Online Net to Target Net

softSync :: Parameterized f => Tensor -> f -> f -> IO f

Softly copy parameters from Online Net to Target Net

copySync :: Parameterized f => f -> f -> f

Hard Copy of Parameter from one net to the other