MLFlow.Extensions

Documentation

sanatizeJSON :: Char -> Char

Sanatize JSON for MLFlow: Names may only contain alphanumerics, underscores (_), dashes (-), periods (.), spaces ( ), and slashes (/).

data Tracker

Data Logging to MLFlow Trackign Server

Constructors

Tracker

Fields

uri :: TrackingURI

Tracking Server URI

experimentId :: ExperimentID

Experiment ID

experimentName :: String

Experiment Name

runIds :: Map String RunID

Run IDs

runId :: Tracker -> String -> RunID

Retrieve a run ID

mkTracker :: TrackingURI -> String -> IO Tracker

Make new Tracker given a Tracking Server URI

mkTracker' :: String -> Int -> String -> IO Tracker

Make new Tracker given a Hostname and Port

newExperiment :: Tracker -> String -> IO Tracker

Create a new Experiment with rng suffix

newExperiment' :: Tracker -> String -> IO Tracker

Create a new Experiment

newRuns :: Tracker -> [String] -> [Param] -> IO Tracker

Create a new run with a set of given paramters

newRuns' :: Int -> Tracker -> IO Tracker

New run with algorithm id and #envs as log params

endRun :: String -> Tracker -> IO Tracker

End a run

endRuns :: Tracker -> IO Tracker

End all runs of a Tracker

endRuns' :: Tracker -> IO ()

End all runs and discard tracker

trackLoss :: Tracker -> Int -> String -> Float -> IO (Response ByteString)

Write Loss to Tracking Server

trackReward :: Tracker -> Int -> Tensor -> IO ()

Write Reward to Tracking Server

sanatizeMap :: Map Int (Map String Float) -> Map Int (Map String Float)

Clean up a Map returned by Circus Server

trackEnvState :: Tracker -> CircusUrl -> Int -> IO ()

Write Current state of the Environment to Trackign Server