RPB

Description

General Replay Buffer Types and TypeClasses

Documentation

data ReplayMemory

Available Replay Buffer Types

Constructors

RPB
Vanilla Replay Buffer

PER
Prioritized Experience Replay

MEM
PPO Style replay Memory

ERE
Emphasizing Recent Experience

HER
Hindsight Experience Replay

class Functor b => ReplayBuffer b where

Replay Buffer Interface

Methods

size :: b Tensor -> Int

Return size of current buffer

push :: Int -> b Tensor -> b Tensor -> b Tensor

Push one buffer into another

lookUp :: [Int] -> b Tensor -> b Tensor

Look Up given list if indices

sampleIO :: Int -> b Tensor -> IO (b Tensor)

Take n Random Samples

asTuple :: b Tensor -> (Tensor, Tensor, Tensor, Tensor, Tensor)

Return the Tuple: (s, a, r, s', d) for training

collectExperience :: Agent a => CircusUrl -> Tracker -> Int -> a -> IO (b Tensor)

Collect Experiences in Buffer

randomBatches :: ReplayBuffer b => Int -> Int -> b Tensor -> IO [Transition]

Generate a list of uniformly sampled minibatches

data Buffer a

Vanilla Replay Buffer

Constructors

Buffer

Fields

states :: !a

States

actions :: !a

Actions

rewards :: !a

Rewards

states' :: !a

Next States

dones :: !a

Terminal Mask

empty :: Buffer Tensor

Create a new, empty Buffer on the CPU

size' :: Buffer Tensor -> Int

How many Trajectories are currently stored in memory

drop :: Int -> Buffer Tensor -> Buffer Tensor

Drop number of entries from the beginning of the Buffer

push' :: Int -> Buffer Tensor -> Buffer Tensor -> Buffer Tensor

Push one buffer into another one

lookUp' :: [Int] -> Buffer Tensor -> Buffer Tensor

Get the given indices from Buffer

sampleIO' :: Int -> Buffer Tensor -> IO (Buffer Tensor)

Take n random samples from Buffer

asTuple' :: Buffer Tensor -> Transition

Return (State, Action, Reward, Next State, Done) Tuple

collectStep :: Agent a => CircusUrl -> Tracker -> Int -> Int -> a -> Tensor -> Buffer Tensor -> IO (Buffer Tensor)

Evaluate Policy for T steps and return experience Buffer

collectExperience' :: Agent a => CircusUrl -> Tracker -> Int -> a -> IO (Buffer Tensor)

Collect experience for a given number of steps