RPB.HER

Description

Hindsight Experience Replay

Documentation

data Strategy

Hindsight Experience Replay Strategies for choosing Goals

Constructors

Final
Only Final States are additional targets

Random
Replay with k random states encountered so far (basically vanilla)

Episode
Replay with k random states from same episode.

Future
Replay with k random states from same episode, that were observed after

data Buffer a

Strict Simple/Naive Replay Buffer

Constructors

Buffer

Fields

states :: !a

States

actions :: !a

Actions

rewards :: !a

Rewards

states' :: !a

Next States

dones :: !a

Terminal Mask

goals :: !a

Desired Goal

goals' :: !a

Acheived Goal

empty :: Buffer Tensor

Construct an empty HER Buffer

envSplit :: Int -> Buffer Tensor -> [Buffer Tensor]

Split buffer collected from pool by env

epsSplit :: Buffer Tensor -> [Buffer Tensor]

Split a buffer into episodes, dropping the last unfinished

sampleGoals :: CircusUrl -> Strategy -> Int -> Buffer Tensor -> IO (Buffer Tensor)

Sample Additional Goals according to Strategy (drop first). Random is basically the same as Episode you just have to give it the entire buffer, not just the episode.