Lib

Description

Utility and Helper functions for EDELWACE

Documentation

data Mode

Run Mode

Constructors

Train
Start Agent Training

Continue
Continue Agent Training

Evaluate
Evaluate Agent

data Args

Command Line Arguments

Constructors

Args

Fields

algorithm :: String

See ALG.Algorithm

memory :: String

See RPB.ReplayMemory

cktHost :: String

Circus Server Host Address

cktPort :: String

Circus Server Port

ace :: String

ACE ID

pdk :: String

ACE PDK

space :: String

Design / Action Space

var :: String

(Non-)Goal Environment

cpPath :: String

Checkpoint Base Path

mlfHost :: String

MLFlow Server Host Address

mlfPort :: String

MLFlow Server Port

mode :: String

Run Mode

type Transition = (Tensor, Tensor, Tensor, Tensor, Tensor)

Type Alias for Transition Tuple (state, action, reward, state', done)

range :: Int -> [Int]

Range from 0 to n - 1

fst' :: (a, b, c) -> a

First of triple

delete' :: Ord a => [a] -> Set a -> Set a

Delete multiple elements from a Set

splits :: [Int] -> [[Int]]

Helper function creating split indices

fst3 :: (a, b, c) -> a

First of Triple

both :: (a -> b) -> (a, a) -> (b, b)

Apply a function to both elements of a Tuple

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

Uncurry Triple

lookup' :: Ord k => [k] -> Map k a -> Maybe [a]

Like lookup but for a list of keys.

tmap :: (Tensor -> Tensor) -> Transition -> Transition

Map an appropriate function over a transition tuple

currentTimeStamp :: String -> IO String

Current Timestamp as formatted string

currentTimeStamp' :: IO String

Current Timestamp with default formatting: "%Y%m%d-%H%M%S"

createModelArchiveDir :: String -> IO String

Create a model archive directory for the given algorithm

createModelArchiveDir' :: String -> String -> String -> String -> String -> String -> IO String

Create a model archive directory for the given algorithm, ace id and backend

saveOptim :: Adam -> FilePath -> IO ()

Optimizer moments at given prefix

loadOptim :: Int -> Float -> Float -> FilePath -> IO Adam

Load Optimizer State

weightLimit :: Linear -> Float

Calculate weight Limits based on Layer Dimensions

data Initializer

Type of weight initialization

Constructors

Normal
Normally distributed weights

Uniform
Uniformally distributed weights

XavierNormal
Using T.xavierNormal

XavierUniform
Using T.xavierUniform

KaimingNormal
Using T.kaimingNormal

KaimingUniform
Using T.kaimingUniform

Dirac

Eye

Ones

Zeros

Constant

initWeights :: Initializer -> Float -> Float -> [Int] -> IO IndependentTensor

Weights for a layer given limits and dimensions.

weightInit :: Initializer -> Float -> Float -> Linear -> IO Linear

Initialize Weights of Linear Layer

weightInit' :: Initializer -> Float -> Float -> Linear -> IO Linear

Initialize Weights and Bias of Linear Layer

weightInitUniform :: Float -> Float -> Linear -> IO Linear

Initialize weights uniformally given upper and lower bounds

weightInitUniform' :: Linear -> IO Linear

Initialize weights uniformally based on Fan In

weightInitNormal :: Float -> Float -> Linear -> IO Linear

Initialize weights normally given mean and std bounds

weightInitNormal' :: Linear -> IO Linear

Initialize weights normally based on Fan In

softUpdate :: Tensor -> Tensor -> Tensor -> Tensor

Softly update parameters from Online Net to Target Net

softSync :: Parameterized f => Tensor -> f -> f -> IO f

Softly copy parameters from Online Net to Target Net

copySync :: Parameterized f => f -> f -> f

Hard Copy of Parameter from one net to the other